pyproject.toml
[tool.poetry]
name = "rss-fetcher-py"
version = "0.1.0"
description = ""
authors = ["Manik <manik6418@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Pyrogram = "^2.0.106"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"
tgcrypto = "^1.2.5"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
